refactor[#17] : Radio,RadioGroup cva 설정
