+++ beta-reader/app/globals.css
@@ -12,10 +12,10 @@
 
   --color-secondary-white: #ffffff;
   --color-secondary-black: #121212;
-  --color-secondary-100: ##d9deee;
-  --color-secondary-200: ##bfc3d3;
-  --color-secondary-300: ##989daf;
-  --color-secondary-400: ##7b8091;
+  --color-secondary-100: #d9deee;
+  --color-secondary-200: #bfc3d3;
+  --color-secondary-300: #989daf;
+  --color-secondary-400: #7b8091;
   --color-secondary-500: #606575;
   --color-secondary-600: #4a4e5b;
   --color-secondary-700: #3a3d46;
