+++ beta-reader/src/shared/ui/Form/ui/Radio.stories.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import { Radio, RadioGroup } from "./Form";
+import { Radio, RadioGroup } from "./Radio";
 import { action } from "@storybook/addon-actions";
 import type { Meta, StoryObj } from "@storybook/react";
